+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.2.0] - 2026-06-14
 
 ### Added
 
@@ -421,7 +421,8 @@ Combining a deprecated alias with its replacement (e.g.
 
 - Initial release.
 
-[Unreleased]: https://github.com/izumo-m/asana-api-cli/compare/v3.1.3...HEAD
+[Unreleased]: https://github.com/izumo-m/asana-api-cli/compare/v3.2.0...HEAD
+[3.2.0]: https://github.com/izumo-m/asana-api-cli/compare/v3.1.3...v3.2.0
 [3.1.3]: https://github.com/izumo-m/asana-api-cli/compare/v3.1.2...v3.1.3
 [3.1.2]: https://github.com/izumo-m/asana-api-cli/compare/v3.1.1...v3.1.2
 [3.1.1]: https://github.com/izumo-m/asana-api-cli/compare/v3.1.0...v3.1.1